requirements.txt
django-money
django-filter
djangorestframework
django-allauth
django_extensions
django-sslserver
boto3
django-storages
django-countries
django-rest-auth
